+++ projects/igniteui-angular/src/lib/grids/grid-base.directive.ts
@@ -836,16 +836,16 @@ export abstract class IgxGridBaseDirective extends DisplayDensityBase implements
     public columnResized = new EventEmitter<IColumnResizeEventArgs>();
 
     /**
-     * Emitted when a cell is right clicked.
+     * Emitted when a cell or row is right clicked.
      *
      * @remarks
-     * Returns the `IgxGridCell` object.
+     * Returns the `IgxGridCell` object if the immediate context menu target is a cell or an `IgxGridRow` otherwise.
      * ```html
      * <igx-grid #grid [data]="localData" (contextMenu)="contextMenu($event)" [autoGenerate]="true"></igx-grid>
      * ```
      */
     @Output()
-    public contextMenu = new EventEmitter<IGridCellEventArgs>();
+    public contextMenu = new EventEmitter<IGridCellEventArgs | IGridRowEventArgs>();
 
     /**
      * Emitted when a cell is double clicked.