+++ projects/igniteui-angular/src/lib/grids/grid/grid.component.spec.ts
@@ -26,7 +26,7 @@ import { ISortingExpression, SortingDirection } from '../../data-operations/sort
 import { GRID_SCROLL_CLASS } from '../../test-utils/grid-functions.spec';
 import { AsyncPipe, NgFor, NgIf } from '@angular/common';
 import { IgxPaginatorComponent, IgxPaginatorContentDirective } from '../../paginator/paginator.component';
-import { IGridRowEventArgs, IgxGridFooterComponent, IgxGridRow, IgxGroupByRow, IgxSummaryRow } from '../public_api';
+import { IGridCellEventArgs, IGridRowEventArgs, IgxGridFooterComponent, IgxGridRow, IgxGroupByRow, IgxSummaryRow } from '../public_api';
 import { getComponentSize } from '../../core/utils';
 
 
@@ -2082,7 +2082,35 @@ describe('IgxGrid Component Tests #grid', () => {
             fix.detectChanges();
             expect(grid.rowClick.emit).toHaveBeenCalledTimes(2);
             expect(grid.rowClick.emit).toHaveBeenCalledWith(args);
-        })
+        });
+
+        fit('Should emit contextMenu when clicking outside of the columns area', () => {
+            const fix = TestBed.createComponent(IgxGridDefaultRenderingComponent);
+            fix.componentInstance.initColumnsRows(5, 5);
+            //fix.componentInstance.columns.forEach(c => c.width = '100px');
+            fix.componentInstance.grid.width = '900px';
+            fix.detectChanges();
+            const grid = fix.componentInstance.grid;
+            grid.columnList.forEach(c => c.width = '100px');
+            fix.detectChanges();
+            const spy = spyOn(grid.contextMenu, 'emit').and.callThrough();
+            const event = new Event('contextmenu');
+            const row = grid.rowList.get(0);
+            const cell = row.cells.get(0);
+            cell.nativeElement.dispatchEvent(event);
+            fix.detectChanges();
+            expect(grid.contextMenu.emit).toHaveBeenCalledTimes(1);
+            expect(grid.contextMenu.emit).toHaveBeenCalledWith(jasmine.objectContaining({
+                cell: jasmine.anything()
+            }));
+            spy.calls.reset();
+            row.nativeElement.dispatchEvent(event);
+            fix.detectChanges();
+            expect(grid.contextMenu.emit).toHaveBeenCalledTimes(1);
+            expect(grid.contextMenu.emit).toHaveBeenCalledWith(jasmine.objectContaining({
+                row: jasmine.anything()
+            }));
+        });
 
         it(`Verify that getRowData returns correct data`, () => {
             const fix = TestBed.createComponent(IgxGridDefaultRenderingComponent);