+++ projects/igniteui-angular/src/lib/grids/row.directive.ts
@@ -415,6 +415,18 @@ export class IgxRowDirective implements DoCheck, AfterViewInit, OnDestroy {
      * @hidden
      * @internal
      */
+    @HostListener('contextmenu', ['$event'])
+    public onContextMenu(event: MouseEvent) {
+        this.grid.contextMenu.emit({
+            row: this,
+            event
+        });
+    }
+
+    /**
+     * @hidden
+     * @internal
+     */
     @HostListener('mouseenter')
     public showActionStrip() {
         if (this.grid.actionStrip) {